+++ src/Framework/src/Application/AphiriaComponents.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Application;
 
-use Aphiria\Application\Builders\IApplicationBuilder;
+use Aphiria\Application\IApplicationBuilder;
 use Aphiria\Application\IComponent;
 use Aphiria\Application\IModule;
 use Aphiria\Authentication\AuthenticationScheme;
@@ -47,8 +47,8 @@ use Aphiria\Framework\Validation\Components\ValidationComponent;
 use Aphiria\Middleware\MiddlewareBinding;
 use Aphiria\Middleware\MiddlewareCollection;
 use Aphiria\Net\Http\HttpStatusCode;
-use Aphiria\Routing\Builders\RouteCollectionBuilder;
-use Aphiria\Validation\Builders\ObjectConstraintsRegistryBuilder;
+use Aphiria\Routing\RouteCollectionBuilder;
+use Aphiria\Validation\ObjectConstraintsRegistryBuilder;
 use Closure;
 use Exception;
 use RuntimeException;
@@ -317,7 +317,6 @@ trait AphiriaComponents
             $appBuilder->withComponent(new ExceptionHandlerComponent(Container::$globalInstance));
         }
 
-        /** @psalm-suppress PossiblyInvalidArgument https://github.com/vimeo/psalm/issues/9747 - bug */
         $appBuilder->getComponent(ExceptionHandlerComponent::class)
             ->withConsoleOutputWriter($exceptionType, $callback);
 
@@ -425,7 +424,6 @@ trait AphiriaComponents
             $appBuilder->withComponent(new ExceptionHandlerComponent(Container::$globalInstance));
         }
 
-        /** @psalm-suppress InvalidArgument https://github.com/vimeo/psalm/issues/9747 - bug */
         $appBuilder->getComponent(ExceptionHandlerComponent::class)
             ->withLogLevelFactory($exceptionType, $logLevelFactory);
 
@@ -511,7 +509,6 @@ trait AphiriaComponents
             $appBuilder->withComponent(new ExceptionHandlerComponent(Container::$globalInstance));
         }
 
-        /** @psalm-suppress PossiblyInvalidArgument https://github.com/vimeo/psalm/issues/9747 - bug */
         $appBuilder->getComponent(ExceptionHandlerComponent::class)
             ->withProblemDetails($exceptionType, $type, $title, $detail, $status, $instance, $extensions);