+++ src/Router/tests/RouteCollectionBuilderTest.php
@@ -10,14 +10,14 @@
 
 declare(strict_types=1);
 
-namespace Aphiria\Routing\Tests\Builders;
+namespace Aphiria\Routing\Tests;
 
-use Aphiria\Routing\Builders\RouteBuilder;
-use Aphiria\Routing\Builders\RouteCollectionBuilder;
-use Aphiria\Routing\Builders\RouteGroupOptions;
 use Aphiria\Routing\Matchers\Constraints\HttpMethodRouteConstraint;
 use Aphiria\Routing\Matchers\Constraints\IRouteConstraint;
 use Aphiria\Routing\Middleware\MiddlewareBinding;
+use Aphiria\Routing\RouteBuilder;
+use Aphiria\Routing\RouteCollectionBuilder;
+use Aphiria\Routing\RouteGroupOptions;
 use PHPUnit\Framework\TestCase;
 
 class RouteCollectionBuilderTest extends TestCase